moved system prg to the R.C.Random module.
- cleaned up the cabal file as a result.
